Changed more wonky pointer casts to uintptr_t.

+++ src/render_soft/soft_d_polyse.c
@@ -117,7 +117,7 @@ void D_PolysetDraw(void) {
     spanpackage_t spans[DPS_MAXSPANS + 1 + ((CACHE_SIZE - 1) / sizeof(spanpackage_t)) + 1];
     // one extra because of cache line pretouching
 
-    a_spans = (spanpackage_t *)(((long)&spans[0] + CACHE_SIZE - 1) & ~(CACHE_SIZE - 1));
+    a_spans = (spanpackage_t *)(((uintptr_t)&spans[0] + CACHE_SIZE - 1) & ~(CACHE_SIZE - 1));
 
     if(r_affinetridesc.drawtype) {
         D_DrawSubdiv();
