+++ src/render_soft/soft_r_main.c
@@ -792,11 +792,11 @@ void R_EdgeDrawing(void) {
     if(auxedges) {
         r_edges = auxedges;
     } else {
-        r_edges = (edge_t *)(((long)&ledges[0] + CACHE_SIZE - 1) & ~(CACHE_SIZE - 1));
+        r_edges = (edge_t *)(((uintptr_t)&ledges[0] + CACHE_SIZE - 1) & ~(CACHE_SIZE - 1));
     }
 
     if(r_surfsonstack) {
-        surfaces = (surf_t *)(((long)&lsurfs[0] + CACHE_SIZE - 1) & ~(CACHE_SIZE - 1));
+        surfaces = (surf_t *)(((uintptr_t)&lsurfs[0] + CACHE_SIZE - 1) & ~(CACHE_SIZE - 1));
         surf_max = &surfaces[r_cnumsurfs];
         // surface 0 doesn't really exist; it's just a dummy because index 0
         // is used to indicate no edge attached to surface
@@ -938,11 +938,11 @@ void R_RenderView(void) {
         Sys_Error("Hunk is missaligned");
     }
 
-    if((long)(&dummy) & 3) {
+    if((uintptr_t)(&dummy) & 3) {
         Sys_Error("Stack is missaligned");
     }
 
-    if((long)(&r_warpbuffer) & 3) {
+    if((uintptr_t)(&r_warpbuffer) & 3) {
         Sys_Error("Globals are missaligned");
     }
